Apply suggestion from review
