+++ lib/modules/versioning/common.ts
@@ -5,3 +5,17 @@ export function isVersioningApiConstructor(
 ): obj is VersioningApiConstructor {
   return typeof obj === 'function';
 }
+
+export function getExcludedVersions(range: string): string[] {
+  return range
+    .split(',')
+    .filter((version) => version.match('!='))
+    .map((version) => version.replace('!=', '').trim());
+}
+
+export function getFilteredRange(range: string): string {
+  return range
+    .split(',')
+    .filter((version) => !version.match('!='))
+    .join(',');
+}
